Delete setup and utilities, move to runtests.jl.
